requirements.txt
torch>=2.1
torchvision
transformers>=4.40
datasets>=2.18
faiss-cpu>=1.7.4
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pandas>=2.0
Pillow>=10.0
tqdm
pytest
